requirements.txt
annotated-types==0.7.0
blinker==1.9.0
blis==1.3.0
cachelib==0.13.0
catalogue==2.0.10
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.3
cymem==2.0.11
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fuzzywuzzy==0.18.0
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Levenshtein==0.27.3
lxml==6.0.0
MarkupSafe==3.0.2
msgspec==0.19.0
murmurhash==1.0.13
mysql-connector==2.2.9
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
preshed==3.0.10
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
PyMySQL==1.1.1
pyOpenSSL==25.3.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.3
pytz==2025.2
RapidFuzz==3.14.3
requests==2.32.4
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
spacy==3.8.8
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
thinc==8.3.8
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.2
Werkzeug==3.1.3
wrapt==2.0.1
